Move db code into repository + other simplifications
